MVP release!
This doesn't do lots of the clever things that RSpec does yet (like
memoizing `let` variables, or letting you define a variable via a
function), but it already scratches an itch that's been bugging me -
reassigning variables within the scope of a nested `describe` block.

This works with a bit of a trick - we use a `beforeAll` hook in each
scope to set up the stack of contexts, and a `beforeEach` to prepare the
context for each test by squashing the context stack down into a single
object.